separate album / photo modules
