+++ resources/js/components/gallery/albumModule/PhotoThumbPanel.vue
