+++ resources/js/components/gallery/albumModule/PhotoThumbPanelList.vue
