+++ resources/js/components/gallery/albumModule/SensitiveWarning.vue
