+++ resources/js/components/gallery/albumModule/thumbs/AlbumThumbOverlay.vue
