+++ resources/js/components/gallery/photoModule/DockButton.vue
