+++ resources/js/components/gallery/photoModule/MapInclude.vue
