+++ resources/js/components/gallery/photoModule/NextPrevious.vue
