+++ resources/js/components/gallery/photoModule/Overlay.vue
