+++ resources/js/views/gallery-panels/Album.vue
@@ -184,10 +184,10 @@
 import { useAuthStore } from "@/stores/Auth";
 import { computed, ref, watch, onMounted, onUnmounted } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import AlbumThumbPanel from "@/components/gallery/AlbumThumbPanel.vue";
-import PhotoThumbPanel from "@/components/gallery/PhotoThumbPanel.vue";
+import AlbumThumbPanel from "@/components/gallery/albumModule/AlbumThumbPanel.vue";
+import PhotoThumbPanel from "@/components/gallery/albumModule/PhotoThumbPanel.vue";
 import ShareAlbum from "@/components/modals/ShareAlbum.vue";
-import AlbumHero from "@/components/gallery/AlbumHero.vue";
+import AlbumHero from "@/components/gallery/albumModule/AlbumHero.vue";
 import AlbumEdit from "@/components/drawers/AlbumEdit.vue";
 import AlbumHeader from "@/components/headers/AlbumHeader.vue";
 import { useLycheeStateStore } from "@/stores/LycheeState";
@@ -202,7 +202,7 @@ import { useAlbumRefresher } from "@/composables/album/albumRefresher";
 import { useContextMenu } from "@/composables/contextMenus/contextMenu";
 import PhotoService from "@/services/photo-service";
 import AlbumService from "@/services/album-service";
-import { AlbumThumbConfig } from "@/components/gallery/thumbs/AlbumThumb.vue";
+import { AlbumThumbConfig } from "@/components/gallery/albumModule/thumbs/AlbumThumb.vue";
 import RenameDialog from "@/components/forms/gallery-dialogs/RenameDialog.vue";
 import AlbumMergeDialog from "@/components/forms/gallery-dialogs/AlbumMergeDialog.vue";
 import MoveDialog from "@/components/forms/gallery-dialogs/MoveDialog.vue";
@@ -211,7 +211,7 @@ import { useGalleryModals } from "@/composables/modalsTriggers/galleryModals";
 import PhotoTagDialog from "@/components/forms/photo/PhotoTagDialog.vue";
 import PhotoCopyDialog from "@/components/forms/photo/PhotoCopyDialog.vue";
 import { Collapse } from "vue-collapsed";
-import SensitiveWarning from "@/components/gallery/SensitiveWarning.vue";
+import SensitiveWarning from "@/components/gallery/albumModule/SensitiveWarning.vue";
 import Unlock from "@/components/forms/album/Unlock.vue";
 import LoginModal from "@/components/modals/LoginModal.vue";
 import Button from "primevue/button";
