+++ resources/js/views/gallery-panels/Albums.vue
@@ -129,7 +129,7 @@
 	/>
 </template>
 <script setup lang="ts">
-import AlbumThumbPanel from "@/components/gallery/AlbumThumbPanel.vue";
+import AlbumThumbPanel from "@/components/gallery/albumModule/AlbumThumbPanel.vue";
 import { useAuthStore } from "@/stores/Auth";
 import { computed, ref, onMounted, onUnmounted } from "vue";
 import AlbumsHeader from "@/components/headers/AlbumsHeader.vue";
@@ -142,7 +142,7 @@ import { useSelection } from "@/composables/selections/selections";
 import { useContextMenu } from "@/composables/contextMenus/contextMenu";
 import ContextMenu from "primevue/contextmenu";
 import { useAlbumsRefresher } from "@/composables/album/albumsRefresher";
-import { AlbumThumbConfig } from "@/components/gallery/thumbs/AlbumThumb.vue";
+import { AlbumThumbConfig } from "@/components/gallery/albumModule/thumbs/AlbumThumb.vue";
 import MoveDialog from "@/components/forms/gallery-dialogs/MoveDialog.vue";
 import AlbumMergeDialog from "@/components/forms/gallery-dialogs/AlbumMergeDialog.vue";
 import DeleteDialog from "@/components/forms/gallery-dialogs/DeleteDialog.vue";
